Break apart _react-entry.jsx file into components.
Just some house cleaning so the top-level app can be unittested.
